+++ src/io/mod.rs
@@ -144,6 +144,41 @@ pub trait Database: Sized {
                         .with(("z", DataType::text_values(["Foo".into(), "Bar".into()]))),
                 )
                 .build(),
+            TableBuilder::new()
+                .name("user_table")
+                .size(100)
+                .schema(
+                    Schema::empty()
+                        .with(("id", DataType::integer_interval(0, 100)))
+                        .with(("name", DataType::text()))
+                        .with(("age", DataType::optional(DataType::float_interval(0., 200.))))
+                        .with(("city", DataType::text_values(["Paris".into(), "New-York".into()]))),
+                )
+                .build(),
+            TableBuilder::new()
+                .name("order_table")
+                .size(200)
+                .schema(
+                    Schema::empty()
+                        .with(("id", DataType::integer_interval(0, 100)))
+                        .with(("user_id", DataType::integer_interval(0, 101)))
+                        .with(("description", DataType::text()))
+                        .with(("date", DataType::date_interval(
+                            chrono::NaiveDate::from_ymd_opt(2020, 12, 06).unwrap(),
+                            chrono::NaiveDate::from_ymd_opt(2023, 12, 06).unwrap(),
+                        ))),
+                )
+                .build(),
+            TableBuilder::new()
+                .name("item_table")
+                .size(300)
+                .schema(
+                    Schema::empty()
+                        .with(("order_id", DataType::integer_interval(0,100)))
+                        .with(("item", DataType::text()))
+                        .with(("price", DataType::float_interval(0.,50.))),
+                )
+                .build(),
         ]
     }
     /// A basic test DB